move reconciler tests to reconciler_test.go file
